# sprint5_biblioteca_api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "📚 Biblioteca REST API",
        "description": "API desarrollada por **Jenifer Alvarez** (Sprint 5 - IT Academy).  \r\n *\r\n * Implementa autenticación con **Laravel Passport**, control de acceso por roles  \r\n * y gestión de **Libros**, **Categorías** y **Préstamos**.  \r\n * Respuestas JSON estructuradas para integrarse con frontends o sistemas de IA.",
        "contact": {
            "name": "Jenifer Álvarez",
            "url": "https://github.com/jenifera5"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/books": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Listar todos los libros",
                "description": "Devuelve un listado de todos los libros registrados en la base de datos.",
                "operationId": "f826b893452fdfde70f403cef218a493",
                "responses": {
                    "200": {
                        "description": "Listado de libros obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "titulo": {
                                                "type": "string",
                                                "example": "Dune"
                                            },
                                            "autor": {
                                                "type": "string",
                                                "example": "Frank Herbert"
                                            },
                                            "anio": {
                                                "type": "integer",
                                                "example": 1965
                                            },
                                            "disponibles": {
                                                "type": "integer",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado o token inválido"
                    }
                }
            }
        },
        "/books": {
            "post": {
                "tags": [
                    "Libros"
                ],
                "summary": "Crear un nuevo libro",
                "operationId": "89f7b090d9d078ec66aa5c369cee0662",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "autor",
                                    "anio",
                                    "disponibles"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "La sombra del viento"
                                    },
                                    "autor": {
                                        "type": "string",
                                        "example": "Carlos Ruiz Zafón"
                                    },
                                    "anio": {
                                        "type": "integer",
                                        "example": 2001
                                    },
                                    "disponibles": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Libro creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Libro creado correctamente"
                                        },
                                        "libro": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "titulo": {
                                                    "type": "string",
                                                    "example": "La sombra del viento"
                                                },
                                                "autor": {
                                                    "type": "string",
                                                    "example": "Carlos Ruiz Zafón"
                                                },
                                                "anio": {
                                                    "type": "integer",
                                                    "example": 2001
                                                },
                                                "disponibles": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos o faltantes"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/books/{id}": {
            "put": {
                "tags": [
                    "Libros"
                ],
                "summary": "Actualizar un libro existente",
                "operationId": "17f8c8125a390ddee0608ec21f680428",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "autor",
                                    "anio",
                                    "disponibles"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "El juego del ángel"
                                    },
                                    "autor": {
                                        "type": "string",
                                        "example": "Carlos Ruiz Zafón"
                                    },
                                    "anio": {
                                        "type": "integer",
                                        "example": 2008
                                    },
                                    "disponibles": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Libro actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Libro actualizado correctamente"
                                        },
                                        "libro": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "titulo": {
                                                    "type": "string",
                                                    "example": "El juego del ángel"
                                                },
                                                "autor": {
                                                    "type": "string",
                                                    "example": "Carlos Ruiz Zafón"
                                                },
                                                "anio": {
                                                    "type": "integer",
                                                    "example": 2008
                                                },
                                                "disponibles": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    },
                    "422": {
                        "description": "Datos inválidos o faltantes"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Libros"
                ],
                "summary": "Eliminar un libro por ID",
                "operationId": "1cb68ffe0dce1c858dbb4476b4c98ef4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Libro eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Libro eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/books/search": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Buscar libros por título o autor",
                "operationId": "2504c05544e200ce6ac8c2bbf4629d0f",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Término de búsqueda (título o autor)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Cervantes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados obtenidos correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resultados obtenidos correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "titulo": {
                                                        "type": "string",
                                                        "example": "Don Quijote de la Mancha"
                                                    },
                                                    "autor": {
                                                        "type": "string",
                                                        "example": "Miguel de Cervantes"
                                                    },
                                                    "anio": {
                                                        "type": "integer",
                                                        "example": 1605
                                                    },
                                                    "disponibles": {
                                                        "type": "integer",
                                                        "example": 4
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/books/stats/popular": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Obtener los 5 libros más prestados",
                "operationId": "86e56f4a0df7030a021b0b1c3e5288f4",
                "responses": {
                    "200": {
                        "description": "Top 5 libros más prestados obtenidos correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Top 5 libros más prestados obtenidos correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "titulo": {
                                                        "type": "string",
                                                        "example": "Cien años de soledad"
                                                    },
                                                    "autor": {
                                                        "type": "string",
                                                        "example": "Gabriel García Márquez"
                                                    },
                                                    "anio": {
                                                        "type": "integer",
                                                        "example": 1967
                                                    },
                                                    "disponibles": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "prestamos_count": {
                                                        "type": "integer",
                                                        "example": 15
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        [
            {
                "bearerAuth": []
            }
        ]
    ]
}